Reference

The following summarizes the documentation of types and methods provided by the TwoDimensional package. This information is also available from the REPL by typing ? followed by the name of a method or a type.

Affine 2D Coordinate Transforms

TwoDimensional.AffineTransforms.AffineTransform — Type

AffineTransform(Axx, Axy, Ax, Ayx, Ayy, Ay) -> A

yields a callable object implementing a 2-dimensional affine transform such that:

A(x, y) -> (Axx*x + Axy*y + Ax, Ayx*x + Ayy*y + Ay)
A((x, y)) -> (Axx*x + Axy*y + Ax, Ayx*x + Ayy*y + Ay)
A(pnt::Point) -> Point(A(pnt.x, pnt.y))
A(pnt::CartesianIndex{2}) -> Point(A(pnt[1], pnt[2]))

The constructor takes one or three parameters:

AffineTransform{T}(Axx, Axy, Ax, Ayx, Ayy, Ay)
AffineTransform{T,R,S}(Axx, Axy, Ax, Ayx, Ayy, Ay)

where T is the concrete floating-point type of the coefficients (Float64 by default), R is the type for storing the factors Axx, Axy, Ayx, and Ayy, and S is the type for storing the offsets Ax and Ay. The bare type of R and S must be T but they may have units.

Changing the floating-point type of an existing 2-dimensional affine transform A can be done by one of:

B = AffineTransform{T}(A)
B = convert(AffineTransform{T}, A)
B = convert_bare_type(T, A)
B = convert_real_type(T, A)
B = convert_floating_point_type(T, A)

the 3 last assume using Unitless. Using Unitless API, the floating-point type T can be retrieved by either of bare_type(A), realtype(A), orfloatingpoint_type(A)withA` a 2-dimensional affine transform instance or type.

Applying the 2-dimensional affine transform A can be done by:

(xp, yp) = A(x,y)       # apply affine transform A to coordinates (x,y)
(xp, yp) = A*(x,y)      # idem
(xp, yp) = A((x,y))     # idem, with pnt = (x,y)
(xp, yp) = A*(x,y)      # idem

A(Point(x,y)) -> Point(xp, yp)
A*Point(x,y)  -> Point(xp, yp)
A⋅Point(x,y)  -> Point(xp, yp)

C = compose(A, B, ...)  # compose 2 (or more) transforms, C = apply B then A
C = A∘B                 # idem
C = A*B                 # idem
C = A⋅B                 # idem

B = translate(x, y, A)  # B = apply A then translate by (x,y)
B = translate(pnt, A)   # idem with pnt = (x,y)
B = pnt + A             # idem

B = translate(A, x, y)  # B = translate by (x,y) then apply A
B = translate(A, pnt)   # idem with pnt = (x,y)
B = A + pnt             # idem

B = rotate(θ, A)   # B = apply A then rotate by angle θ
C = rotate(A, θ)   # C = rotate by angle θ then apply A

B = scale(ρ, A)    # B = apply A then scale by ρ
B = ρ*A            # idem
C = scale(A, ρ)    # C = scale by ρ then apply A
C = A*ρ            # idem

B = inv(A)         # reciprocal coordinate transform
C = A/B            # right division, same as: C = compose(A, inv(B))
C = A\B            # left division, same as: C = compose(inv(A), B)

"∘" and "⋅" can be typed by \circ<tab> and \cdot<tab>.

TwoDimensional.AffineTransforms.scale — Function

B = scale(ρ, A)
B = ρ*A
C = scale(A, ρ)
C = A*ρ

yield 2-dmensional affine transforms B and C such that:

B(x,y) -> ρ*A(x,y)
C(x,y) -> A(ρ*x,ρ*y)

See also: AffineTransform, rotate, translate.

TwoDimensional.AffineTransforms.rotate — Function

B = rotate(θ, A)
C = rotate(A, θ)

yield 2-dmensional affine transforms B and C such that:

B(x,y) = (R∘A)(x,y) = R(A(x,y))
C(x,y) = (A∘R)(x,y) = A(R(x,y))

where R implements rotation by angle θ counterclockwise around the origin at coordimates (0,0). The rotation angle θ is assumed to be in radians if it has no units.

See also: AffineTransform, scale, translate.

TwoDimensional.AffineTransforms.translate — Function

B = translate(x, y, A)
B = translate((x,y), A)
B = translate(pnt, A)
B = (x,y) + A
B = pnt + A

perform a left-translation of the 2-dimensional affine transform A. Applying B yields the same result as if coordinates (x,y) are added to the output of A. Here, pnt = Point(x,y) or pnt = CartesianIndex(x,y).

C = translate(A, x, y)
C = translate(A, (x,y))
C = translate(A, pnt)
C = A + (x,y)
C = A + pnt

perform a right-translation of the 2-dimensional affine transform A. Applying B yields the same result as if coordinates (x,y) are added to the input of A.

See also: AffineTransform, rotate, scale.

TwoDimensional.AffineTransforms.jacobian — Function

jacobian(A::TwoDimensional.AffineTransform)

yields the Jacobian of the affine transform A, that is the absolute value of the determinant of its linear part.

TwoDimensional.AffineTransforms.intercept — Function

intercept(A) returns the tuple (x,y) such that A(x,y) = (0,0).

TwoDimensional.AffineTransforms.compose — Function

compose(A::TwoDimensional.AffineTransform, B::TwoDimensional.AffineTransform)

yields the affine transform which combines the two affine transforms A and B, that is the affine transform which applies B and then A. Composition is accessible via: A∘B, A*B or A⋅B ("∘" and "⋅" can be typed by \circ<tab> and \cdot<tab>).

It is possible to compose more than two affine transforms. For instance, compose(A,B,C) yields the affine transform which applies C then B, then A.

Points

TwoDimensional.AbstractPoint — Type

AbstractPoint{T}

is the abstract type of objects with at least 2 properties: x and y, their respective abscissa and ordinate, both of type T.

See also: Point, WeightedPoint.

TwoDimensional.Point — Type

Point(x,y)
Point((x,y))

yield an instance of a 2D point of coordinates (x,y).

A point may be multiplied or divided by a scalar to scale its coordinates. The addition (resp. subtraction) of two points adds (resp. subtracts) their coordinates.

Coordinates can be specified by keywords:

Point(x=xval, y=yval)

There are no default values for keywords x and y so both must be specified.

The coordinates of a Point, say pnt, can be retrieved as follows:

pnt.x  or  pnt[1]  ->  x
pnt.y  or  pnt[2]  ->  y

or:

x, y = pnt

See also: WeightedPoint, AbstractPoint.

TwoDimensional.WeightedPoint — Type

WeightedPoint{T}(w,x,y)

yields a weighted point which has 3 fields: w its weight, x its abscissa and y its ordinate, all of type T. By convention w ≥ 0 but this is not checked for efficiency reasons.

See also: Point, AbstractPoint.

TwoDimensional.distance — Function

distance(A, B)

yields the Euclidean distance between the 2 points A and B.

Bounding-Boxes

TwoDimensional.BoundingBox — Type

BoundingBox(xmin,xmax,ymin,ymax)
BoundingBox((xmin,ymin),(xmax,ymax))

yield an instance of a 2D rectangular bounding-box whose sides are aligned with the coordinate axes and containing points of coordinates (x,y) such that xmin ≤ x ≤ xmax and ymin ≤ y ≤ ymax. The box is empty if xmin > xmax or ymin > ymax.

A bounding-box can be constructed from the first and last points (i.e. at the lower-left and upper right opposite corners) of the box:

BoundingBox(P0::Point, P1::Point)
BoundingBox(I0::CartesianIndex{2}, I1::CartesianIndex{2})

Coordinates can be specified by keywords:

BoundingBox(xmin=x0, ymin=y0, xmax=x1, ymax=y1)

There are no default values for keywords xmin, xmax, ymin and ymax so all must be specified.

The coordinates of a BoundingBox, say box, can be retrieved as follows:

box.xmin  or  box[1]  ->  xmin
box.xmax  or  box[2]  ->  xmax
box.ymin  or  box[3]  ->  ymin
box.ymax  or  box[4]  ->  ymax

or:

xmin, xmax, ymin, ymax = box

See also Point, interior, exterior.

TwoDimensional.area — Function

area(box)

yields the area of the bounding-box box.

TwoDimensional.center — Function

center(box::BoundingBox) -> c::Point

yields the central point of the bounding-box box.

TwoDimensional.interior — Function

interior([T,] box)

yields the largest bounding-box with integer valued bounds and which is contained by the bounding-box box. Optional argument T is to specify the type of the result or of the coordinates of the result which is the same as box by default.

See also: exterior, round.

TwoDimensional.exterior — Function

exterior([T,] box)

yields the smallest bounding-box with integer valued bounds and which contains the bounding-box box. Optional argument T is to specify the type of the result or of the coordinates of the result which is the same as box by default.

See also: interior, round.

Methods

Missing docstring.

Missing docstring for round(::Union{Point,BoundingBox}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for floor(::Point). Check Documenter's build log for details.

Missing docstring.

Missing docstring for ceil(::Point). Check Documenter's build log for details.

Aliases

TwoDimensional.AbstractPoint2D — Type

AbstractPoint2D{T} is an alias for TwoDimensional.AbstractPoint{T}.

TwoDimensional.Point2D — Type

Point2D{T} is an alias for TwoDimensional.Point{T}.

TwoDimensional.WeightedPoint2D — Type

WeightedPoint2D{T} is an alias for TwoDimensional.WeightedPoint{T}.

TwoDimensional.BoundingBox2D — Type

BoundingBox2D{T} is an alias for TwoDimensional.BoundingBox{T}.

TwoDimensional.AffineTransform2D — Type

AffineTransform2D{T} is an alias for TwoDimensional.AffineTransform{T}.