Reference

The following summarizes the documentation of types and methods provided by the TwoDimensional package. This information is also available from the REPL by typing ? followed by the name of a method or a type.

Points

TwoDimensional.AbstractPoint — Type

AbstractPoint{T} <: TwoDimensional.GeometricElement{T}

is the abstract type of objects with at least 2 properties: x and y, their respective abscissa and ordinate, both of type T.

See also Point.

TwoDimensional.AbstractPoint2D — Type

TwoDimensional.AbstractPoint2D{T}

is an alias for TwoDimensional.AbstractPoint{T}.

TwoDimensional.Point — Type

pnt = Point{T}(x,y)
pnt = Point{T}((x,y))
pnt = Point{T}(; x=..., y=...)
pnt = Point{T}(; r=..., θ=...)

construct a point given its Cartesian coordinates (x,y) (in the 3 first examples) or its polar coordinates (in the last example) with r the distance to the origin and θ the counterclockwise angle relative to x-axis. Parameter T is the type used to store coordinates, it may be omitted.

Note that, in TwoDimensional, x and y respectively correspond to the 1st and 2nd dimensions of 2-dimensional arrays.

A point may be multiplied or divided by a scalar to scale its coordinates. The addition (resp. subtraction) of two points adds (resp. subtracts) their coordinates.

Points have the following properties which reflect the keywords accepted by their constructor:

pnt.x  ->  x::T
pnt.y  ->  y::T

Points are indexable iterators:

length(pnt)     -> 2
firstindex(pnt) -> 1
lastindex(pnt)  -> 2
pnt[1]          -> pnt.x
pnt[2]          -> pnt.y
first(pnt)      -> pnt.x
last(pnt)       -> pnt.y
Tuple(pnt)      -> (pnt.x, pnt.y)

The coordinates of pnt can thus be retrieved by:

x, y = pnt

the polar coordinates of the point can be retrieved by hypot(pnt) -> r, abs(pnt) -> r, or norm(pnt) -> r, and by atan(pnt) -> θ.

See also AbstractPoint, Rectangle, and BoundingBox.

TwoDimensional.Point2D — Type

TwoDimensional.Point2D{T}

is an alias for TwoDimensional.Point{T}.

TwoDimensional.distance — Function

distance(A, B)

yields the Euclidean distance between the 2 points A and B.

Bounding-Boxes

TwoDimensional.BoundingBox — Type

box = BoundingBox{T}(start, stop)
box = BoundingBox{T}((xmin,ymin), (xmax,ymax))
box = BoundingBox{T}(; start=..., stop=...)
box = BoundingBox{T}(; xmin=..., xmax=..., ymin=..., ymax=...)

construct a rectangular bounding-box with edges aligned with the Cartesian axes and given the coordinates of 2 opposite corners, start and stop, whose coordinates, (xmin,ymin) and (xmax,ymax), may be specified as points, as 2-tuples, as 2-dimensional Cartesian indices, or by keywords. Parameter T is the type used to store coordinates, it may be omitted.

Bounding-boxes have the following properties which reflect the keywords accepted by their constructor:

box.xmin  -> xmin::T
box.xmax  -> xmax::T
box.ymin  -> ymin::T
box.ymax  -> ymax::T
box.start -> start::Point{T}
box.stop  -> stop::Point{T}

A bounding-box is assumed to contain all points of coordinates (x,y) such that xmin ≤ x ≤ xmax and ymin ≤ y ≤ ymax. If xmin > xmax or ymin > ymax, the bounding-box is considered as empty.

Boxes are used to represent grid cells and bounding-boxes of other geometric shape. Use TwoDimensional.Rectangle if you want to define rectangular masks.

Bounding-boxes are indexable iterators:

length(box)     -> 2
firstindex(box) -> 1
lastindex(box)  -> 2
box[1]          -> box.start
box[2]          -> box.stop
first(box)      -> box.start
last(box)       -> box.stop
Tuple(box)      -> (box.start, box.stop)

Hence, the parameters of a bounding-box can be retrieved by:

start, stop = box
(xmin, ymin), (xmax, ymax) = box

See also Point, Rectangle, interior, and exterior.

BoundingBox{T}(obj)

yields the bounding-box of the geometric object obj. If the coordinate type T is not provided, T = coord_type(obj) is assumed.

This can be used to compute the union or the intersection of the bounding-boxes of objects:

mapreduce(BoundingBox, ∪, (obj1, obj2, obj3, ...))
mapreduce(BoundingBox, ∩, [obj1, obj2, obj3, ...])

BoundingBox(f, A::AbstractMatrix)

yields the minimal bounding-box of the entries of A such that f(A[i,j]) is true. This can be used to extract this rectangular region:

A[BoundingBox(f, A)]

If A is a matrix of Booleans, f is assumed to be the identity if not specified.

TwoDimensional.BoundingBox2D — Type

TwoDimensional.BoundingBox2D{T}

is an alias for TwoDimensional.BoundingBox{T}.

TwoDimensional.interior — Function

interior([T,] box)

yields the largest bounding-box with integer valued bounds and which is contained by the bounding-box box. Optional argument T is to specify the type of the result or the type of the coordinates of the result which is the same as box by default.

See also exterior, ceil and floor.

TwoDimensional.exterior — Function

exterior([T,] box)

yields the smallest bounding-box with integer valued bounds and which contains the bounding-box box. Optional argument T is to specify the type of the result or the type of the coordinates of the result which is the same as box by default.

See also interior, ceil and floor.

TwoDimensional.grow — Function

TwoDimensional.grow(box, dx, dy=dx)

yields a new bounding-box object corresponding to the input box object with 1st and 2nd dimensions respectively grown by dx and dy.

Note that the algebraic (not absolute) values are applied. Hence, if dx and dy are both negative, the bounding-box is effectively shrunk by abs(dx) and abs(dy).

See also TwoDimensional.shrink.

TwoDimensional.shrink — Function

TwoDimensional.shrink(box, dx, dy=dx)

yields a new bounding-box object corresponding to the input box object with 1st and 2nd dimensions respectively shrunk by dx and dy.

Note that the algebraic (not absolute) values are applied. Hence, if dx and dy are both negative, the bounding-box is effectively grown by abs(dx) and abs(dy).

See also TwoDimensional.grow.

Rectangles

TwoDimensional.Rectangle — Type

rect = Rectangle{T}(start, stop)
rect = Rectangle{T}((x0,y0), (x1,y1))
rect = Rectangle{T}(; start=..., stop=...)
rect = Rectangle{T}(; x0=..., x1=..., y0=..., y1=...)

construct a rectangular rectangle with edges aligned with the Cartesian axes and given the coordinates of 2 opposite corners, start and stop, whose coordinates, (x0,y0) and (x1,y1), may be specified as points, as 2-tuples, as 2-dimensional Cartesian indices, or by keywords. Parameter T is the type used to store coordinates, it may be omitted.

Rectangles have the following properties which reflect the keywords accepted by their constructor:

rect.x0  -> min(x0, x1)::T
rect.x1  -> max(x0, x1)::T
rect.y0  -> min(y0, y1)::T
rect.y1  -> max(y0, y1)::T
rect.start -> start::Point{T}
rect.stop  -> stop::Point{T}

Note that the coordinates are sorted. A rectangle is never empty and contains at least a single point.

Rectangles are indexable iterators:

length(rect)     -> 2
firstindex(rect) -> 1
lastindex(rect)  -> 2
rect[1]          -> rect.start
rect[2]          -> rect.stop
first(rect)      -> rect.start
last(rect)       -> rect.stop
Tuple(rect)      -> (rect.start, rect.stop)

Hence, the parameters of a rectangle can be retrieved by:

start, stop = rect
(x0, y0), (x1, y1) = rect

See also Point, BoundingBox, interior, and exterior.

TwoDimensional.Rectangle2D — Type

TwoDimensional.Rectangle2D{T}

is an alias for TwoDimensional.Rectangle{T}.

TwoDimensional.area — Function

TwoDimensional.area(rect)

yields the area of the rectangle rect.

Coordinate Type

TwoDimensional.coord_type — Function

coord_type(obj) -> T

yields the coordinate type of a geometrical object or of its type.

TwoDimensional.convert_coord_type — Function

convert_coord_type(T::Type, obj::GeometricObject)

converts the coordinate type of a geometrical object obj to T. If the coordinate type of obj is already T, obj itself is returned.

convert_coord_type(T::Type, objs::GeometricObject...) -> objs′

yields the tuple of geometric objects objs... converted to the coordinate type T.

TwoDimensional.promote_coord_type — Function

promote_coord_type(types::Type{<:GeometricObject}...) -> T

yields the promoted coordinate type of geometric object types in types...

promote_coord_type(objs::GeometricObject...) -> objs′

yields the tuple of geometric objects objs... converted to the same coordinate type.

Affine 2D Coordinate Transforms

TwoDimensional.AffineTransform — Type

AffineTransform(Axx, Axy, Ax, Ayx, Ayy, Ay) -> A

yields a callable object implementing a 2-dimensional affine transform such that:

A(x, y) -> (Axx*x + Axy*y + Ax, Ayx*x + Ayy*y + Ay)
A((x, y)) -> (Axx*x + Axy*y + Ax, Ayx*x + Ayy*y + Ay)
A(pnt::Point) -> Point(A(pnt.x, pnt.y))
A(pnt::CartesianIndex{2}) -> Point(A(pnt[1], pnt[2]))

The constructor takes one or three parameters:

AffineTransform{T}(Axx, Axy, Ax, Ayx, Ayy, Ay)
AffineTransform{T,R,S}(Axx, Axy, Ax, Ayx, Ayy, Ay)

where T is the concrete floating-point type of the coefficients (Float64 by default), R is the type for storing the factors Axx, Axy, Ayx, and Ayy, and S is the type for storing the offsets Ax and Ay. The bare type of R and S must be T but they may have units.

Changing the floating-point type of an existing 2-dimensional affine transform A can be done by one of:

B = AffineTransform{T}(A)
B = convert(AffineTransform{T}, A)
B = convert_bare_type(T, A)
B = convert_real_type(T, A)
B = convert_floating_point_type(T, A)

the 3 last assume using Unitless. Using Unitless API, the floating-point type T can be retrieved by either of bare_type(A), realtype(A), orfloatingpoint_type(A)withA` a 2-dimensional affine transform instance or type.

Applying the 2-dimensional affine transform A can be done by:

(xp, yp) = A(x,y)       # apply affine transform A to coordinates (x,y)
(xp, yp) = A*(x,y)      # idem
(xp, yp) = A((x,y))     # idem, with pnt = (x,y)
(xp, yp) = A*(x,y)      # idem

A(Point(x,y)) -> Point(xp, yp)
A*Point(x,y)  -> Point(xp, yp)
A⋅Point(x,y)  -> Point(xp, yp)

C = compose(A, B, ...)  # compose 2 (or more) transforms, C = apply B then A
C = A∘B                 # idem
C = A*B                 # idem
C = A⋅B                 # idem

B = translate(x, y, A)  # B = apply A then translate by (x,y)
B = translate(pnt, A)   # idem with pnt = (x,y)
B = pnt + A             # idem

B = translate(A, x, y)  # B = translate by (x,y) then apply A
B = translate(A, pnt)   # idem with pnt = (x,y)
B = A + pnt             # idem

B = rotate(θ, A)   # B = apply A then rotate by angle θ
C = rotate(A, θ)   # C = rotate by angle θ then apply A

B = scale(ρ, A)    # B = apply A then scale by ρ
B = ρ*A            # idem
C = scale(A, ρ)    # C = scale by ρ then apply A
C = A*ρ            # idem

B = inv(A)         # reciprocal coordinate transform
C = A/B            # right division, same as: C = compose(A, inv(B))
C = A\B            # left division, same as: C = compose(inv(A), B)

"∘" and "⋅" can be typed by \circ<tab> and \cdot<tab>.

AffineTransform() -> Id
AffineTransform{T}() -> Id
AffineTransform{T,R,S}() -> Id

yields a 2-dimensional affine transform corresponding to the identity (up to possible change of type and units). Parameter T is the floating-point type of the coefficients (Float64 by default). Parameters R and S are the types of the factors and of the offsets (by default both are assumed to be T). These are shortcuts to:

 AffineTransform(oneunit(R),zero(R),zero(S), zero(R),oneunit(R),zero(S))

TwoDimensional.AffineTransform2D — Type

TwoDimensional.AffineTransform2D{T,R,S}

is an alias for TwoDimensional.AffineTransform{T,R,S}.

TwoDimensional.compose — Function

compose(A::TwoDimensional.AffineTransform, B::TwoDimensional.AffineTransform)

yields the affine transform which combines the two affine transforms A and B, that is the affine transform which applies B and then A. Composition is accessible via: A∘B, A*B or A⋅B ("∘" and "⋅" can be typed by \circ<tab> and \cdot<tab>).

It is possible to compose more than two affine transforms. For instance, compose(A,B,C) yields the affine transform which applies C then B, then A.

TwoDimensional.intercept — Function

intercept(A)
intercept(T<:Point, A)

return the tuple (x,y) such that A(x,y) = (0,0). If a Point type T is specified, a point of this type with coordinates (x,y) is returned.

TwoDimensional.jacobian — Function

jacobian(A::TwoDimensional.AffineTransform)

yields the Jacobian of the affine transform A, that is the absolute value of the determinant of its linear part.

TwoDimensional.rotate — Function

B = rotate(θ, A)
C = rotate(A, θ)

yield 2-dmensional affine transforms B and C such that:

B(x,y) = (R∘A)(x,y) = R(A(x,y))
C(x,y) = (A∘R)(x,y) = A(R(x,y))

where R implements rotation by angle θ counterclockwise around the origin at coordimates (0,0). The rotation angle θ is assumed to be in radians if it has no units.

See also: AffineTransform, scale, translate.

Methods

TwoDimensional.center — Function

center(obj::GeometricObject) -> c::Point

yields the central point of the geometric object obj.

TwoDimensional.scale — Function

B = scale(ρ, A)
B = ρ*A
C = scale(A, ρ)
C = A*ρ

yield 2-dmensional affine transforms B and C such that:

B(x,y) -> ρ*A(x,y)
C(x,y) -> A(ρ*x,ρ*y)

See also: AffineTransform, rotate, translate.

TwoDimensional.translate — Function

B = translate(x, y, A)
B = translate((x,y), A)
B = translate(pnt, A)
B = (x,y) + A
B = pnt + A

perform a left-translation of the 2-dimensional affine transform A. Applying B yields the same result as if coordinates (x,y) are added to the output of A. Here, pnt = Point(x,y) or pnt = CartesianIndex(x,y).

C = translate(A, x, y)
C = translate(A, (x,y))
C = translate(A, pnt)
C = A + (x,y)
C = A + pnt

perform a right-translation of the 2-dimensional affine transform A. Applying B yields the same result as if coordinates (x,y) are added to the input of A.

See also: AffineTransform, rotate, scale.

Base.floor — Method

floor([T,] obj)

applies the floor function to the coordinates of the vertices defining the vertex-based geometric object obj and returns an object of the same kind built with the resulting vertices. Optional argument T can be the type of the returned object or the type of the coordinates for the returned object.

See also: round, ceil.

For bounding-boxes, see also: interior, exterior.

Base.ceil — Method

ceil([T,] obj)

applies the ceil function to the coordinates of the vertices defining the vertex-based geometric object obj and returns an object of the same kind built with the resulting vertices. Optional argument T can be the type of the returned object or the type of the coordinates for the returned object.

See also: round, floor.

For bounding-boxes, see also: interior, exterior.

Base.round — Method

round([T,] obj, [r::RoundingMode])

applies the round function to the coordinates of the vertices defining the vertex-based geometric object obj and returns an object of the same kind built with the resulting vertices. Optional argument T can be the type of the returned object or the type of the coordinates for the returned object. Rounding mode may be specified by optional argument r, the default being the same as the round method for a scalar value.

See also: floor, ceil.