Reference

The following provides detailled documentation about types and methods provided by the TwoDimensional package. This information is also available from the REPL by typing ? followed by the name of a method or a type.

Affine 2D Coordinate Transforms

TwoDimensional.AffineTransforms.AffineTransform — Type

Affine 2D Transforms

An affine 2D transform A is defined by 6 real coefficients, Axx, Axy, Ax, Ayx, Ayy and Ay. Such a transform maps (x,y) as (xp,yp) given by:

xp = Axx*x + Axy*y + Ax
yp = Ayx*x + Ayy*y + Ay

The immutable type AffineTransform is used to store an affine 2D transform A, it can be created by:

I = AffineTransform{T}() # yields the identity with type T
A = AffineTransform{T}(Axx, Axy, Ax, Ayx, Ayy, Ay)

The parameter T above is used to specify the floating-point type for the coefficients; if omitted it is guessed from the types of the coefficients.

Operations with affine 2D transforms

Many operations are available to manage or apply affine transforms:

(xp, yp) = A(x,y)       # apply affine transform A to coordinates (x,y)
(xp, yp) = A*(x,y)      # idem
(xp, yp) = A(v)         # idem, with v = (x,y)
(xp, yp) = A*v          # idem

A(Point(x,y)) -> Point(xp, yp)
A*Point(x,y)  -> Point(xp, yp)

C = compose(A, B, ...)  # compose 2 (or more) transforms, C = apply B then A
C = A∘B                 # idem
C = A*B                 # idem
C = A⋅B                 # idem

B = translate(x, y, A)  # B = apply A then translate by (x,y)
B = translate(v, A)     # idem with v = (x,y)
B = v + A               # idem

B = translate(A, x, y)  # B = translate by (x,y) then apply A
B = translate(A, v)     # idem with v = (x,y)
B = A + v               # idem

B = rotate(θ, A)   # B = apply A then rotate by angle θ
C = rotate(A, θ)   # C = rotate by angle θ then apply A

B = scale(ρ, A)    # B = apply A then scale by ρ
B = ρ*A            # idem
C = scale(A, ρ)    # C = scale by ρ then apply A
C = A*ρ            # idem

B = inv(A)         # reciprocal coordinate transform
C = A/B            # right division, same as: C = compose(A, inv(B))
C = A\B            # left division, same as: C = compose(inv(A), B)

"∘" and "⋅" can be typed by \circ<tab> and \cdot<tab>.

Type conversion

As a general rule, the floating-point type T of an AffineTransform{T} is imposed for all operations and for the result. The floating-point type of the composition of several coordinate transforms is the promoted type of the transforms which have been composed.

Calling eltype(A) yields floating-point type of the coefficients of the 2D affine transform A. To convert the floating-point type of the coefficients of A to be T, do one of:

B = T.(A)
B = AffineTransform{T}(A)
B = convert(AffineTransform{T}, A)

TwoDimensional.AffineTransforms.scale — Function

Scaling an affine transform

There are two ways to combine a scaling by a factor ρ with an affine transform A. Left-scaling as in:

B = scale(ρ, A)

results in scaling the output of the transform; while right-scaling as in:

C = scale(A, ρ)

results in scaling the input of the transform. The above examples yield transforms which behave as:

B(v) = ρ.*A(v)
C(v) = A(ρ.*v)

where v is any 2-element tuple.

The same results can be obtained with the * operator:

B = ρ*A    # same as: B = scale(ρ, A)
C = A*ρ    # same as: B = scale(A, ρ)

See also: AffineTransform, rotate, translate.

TwoDimensional.AffineTransforms.rotate — Function

Rotating an affine transform

There are two ways to combine a rotation by angle θ (in radians counterclockwise) with an affine transform A. Left-rotating as in:

B = rotate(θ, A)

results in rotating the output of the transform; while right-rotating as in:

C = rotate(A, θ)

results in rotating the input of the transform. The above examples are similar to:

B = R∘A
C = A∘R

where R implements rotation by angle θ around (0,0).

See also: AffineTransform, scale, translate.

TwoDimensional.AffineTransforms.translate — Function

Translating an affine transform

Affine transforms can be letf- or right-translated.

translate(x, y, A)

or

translate((x,y), A)

yield an affine transform which translate the output of affine transform A by offsets x and y.

translate(A, x, y)

or

translate(A, (x,y))

yield an affine transform which translate the input of affine transform A by offsets x and y.

The same results can be obtained with the + operator:

B = (x,y) + A    # same as: B = translate((x,y), A)
B = A + (x,y)    # same as: B = translate(A, (x,y))

See also: AffineTransform, rotate, scale.

TwoDimensional.AffineTransforms.jacobian — Function

jacobian(A) returns the Jacobian of the affine transform A, that is the absolute value of the determinant of its linear part.

TwoDimensional.AffineTransforms.intercept — Function

intercept(A) returns the tuple (x,y) such that A(x,y) = (0,0).

TwoDimensional.AffineTransforms.compose — Function

compose(A,B) yields the affine transform which combines the two affine transforms A and B, that is the affine transform which applies B and then A. Composition is accessible via: A∘B, A*B or A⋅B ("∘" and "⋅" can be typed by \circ<tab> and \cdot<tab>).

It is possible to compose more than two affine transforms. For instance, compose(A,B,C) yields the affine transform which applies C then B, then A.

Points

TwoDimensional.AbstractPoint — Type

Any object whose type is derived from AbstractPoint{T} has at least 2 fields: x its abscissa and y its ordinate, both of type T.

See also: Point, WeightedPoint.

TwoDimensional.Point — Type

Point(x,y)

yields an instance of a 2D point of coordinates (x,y).

A point may be multiplied or divided by a scalar to scale its coordinates. The addition (resp. subtraction) of two points adds (resp. subtracts) their coordinates.

Coordinates can be specified by keywords:

Point(x=xval, y=yval)

There are no default values for keywords x and y so both must be specified.

The coordinates of a Point, say pnt, can be retrieved as follows:

pnt.x  or  pnt[1]  ->  x
pnt.y  or  pnt[2]  ->  y

See also: WeightedPoint, AbstractPoint.

TwoDimensional.WeightedPoint — Type

A WeightedPoint{T} has just 3 fields: w its weight, x its abscissa and y its ordinate, all of type T. By convention w ≥ 0 but this is not checked for efficiency reasons.

See also: Point, AbstractPoint.

TwoDimensional.distance — Function

distance(A, B)

yields the Euclidean distance between the 2 points A and B.

Bounding-Boxes

TwoDimensional.BoundingBox — Type

BoundingBox(xmin,xmax,ymin,ymax) yields an instance of a 2D rectangular bounding-box whose sides are aligned with the coordinate axes and containing points of coordinates (x,y) such that xmin ≤ x ≤ xmax and ymin ≤ y ≤ ymax. The box is empty if xmin > xmax or ymin > ymax.

A bounding-box can be constructed from the first and last points (i.e. at the lower-left and upper right opposite corners) of the box:

BoundingBox(P0::Point, P1::Point)
BoundingBox(I0::CartesianIndex{2}, I1::CartesianIndex{2})

Coordinates can be specified by keywords:

BoundingBox(xmin=x0, ymin=y0, xmax=x1, ymax=y1)

There are no default values for keywords xmin, xmax, ymin and ymax so all must be specified.

The coordinates of a BoundingBox, say box, can be retrieved as follows:

box.xmin  or  box[1]  ->  xmin
box.xmax  or  box[2]  ->  xmax
box.ymin  or  box[3]  ->  ymin
box.ymax  or  box[4]  ->  ymax

See also Point, interior, exterior.

TwoDimensional.area — Function

area(B) yields the area of the bounding-box B.

TwoDimensional.center — Function

center(B::BoundingBox) -> c::Point

yields the central point of the bounding-box B.

TwoDimensional.interior — Function

interior([T,] B)

yields the largest bounding-box with integer valued bounds and which is contained by the bounding-box B. Optional argument T is to specify the type of the result or of the coordinates of the result which is the same as B by default.

See also: exterior, round.

TwoDimensional.exterior — Function

exterior([T,] B)

yields the smallest bounding-box with integer valued bounds and which contains the bounding-box B. Optional argument T is to specify the type of the result or of the coordinates of the result which is the same as B by default.

See also: interior, round.

Methods

Base.round — Method

round([T,] obj)

yields the object that is the nearest to obj by rounding its coordinates to the nearest integer. Argument T can be the type of the result (a point or a bounding-box) or the type of the coordinates of the result.

For points, see also: floor, ceil.

For bounding-boxes, see also: interior, exterior.

Aliases

TwoDimensional.AbstractPoint2D — Type

AffineTransform{T} is an alias for TwoDimensional.AbstractPoint{T}.

TwoDimensional.Point2D — Type

Point2D{T} is an alias for TwoDimensional.Point{T}.

TwoDimensional.WeightedPoint2D — Type

WeightedPoint2D{T} is an alias for TwoDimensional.WeightedPoint{T}.

TwoDimensional.BoundingBox2D — Type

BoundingBox2D{T} is an alias for TwoDimensional.BoundingBox{T}.

TwoDimensional.AffineTransform2D — Type

AffineTransform2D{T} is an alias for TwoDimensional.AffineTransform{T}.